Use the default color that Windows uses for progress bars in the splash screen
Currently the color of the progress bar on the splash window uses #00FF00 which stands out and looks like an obvious design mistake, even if seen for only a few seconds. This change proposes to instead use the color that Windows uses for its own progress bars. It is still green, but less bright, and something that users will be familiar with.